Implementación de mejoras en el front

+++ src/pages/Principal.js
@@ -1,7 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import { Container, Box, TextField, Button, Avatar, Typography, List, ListItem, ListItemText, ListItemAvatar, IconButton, Divider } from '@mui/material';
+import { Container, Box, TextField, Button, Avatar, Typography, List, ListItem, ListItemText, ListItemAvatar, IconButton, Divider, Paper } from '@mui/material';
 import { Search, MoreVert, Delete } from '@mui/icons-material';
 import MarkMesser from '../images/Logo.webp'; // Ruta a tu imagen local
+import PagoFicticioModal from '../components/PagoFicticioModal';
+import chatBackground from '../images/fondoapp.webp';
 
 const Principal = () => {
   const [messages, setMessages] = useState([]);
@@ -12,11 +14,12 @@ const Principal = () => {
   const [inputValue, setInputValue] = useState('');
   const [searchTerm, setSearchTerm] = useState('');
   const [ws, setWs] = useState(null);
+  const [isPaymentModalOpen, setIsPaymentModalOpen] = useState(false);
 
   const fetchUsers = async () => {
     const user = JSON.parse(localStorage.getItem('user'));
     setCurrentUser(user);
-    const response = await fetch(`http://localhost:3000/api/users?userId=${user.id}`);
+    const response = await fetch(`http://localhost:3360/api/users?userId=${user.id}`);
     const data = await response.json();
     const otherUsers = data.filter(u => u.id !== user.id);
     setUsers(otherUsers);
@@ -24,7 +27,7 @@ const Principal = () => {
   };
 
   const fetchMessages = async (userId) => {
-    const response = await fetch(`http://localhost:3000/api/messages/user/${userId}`);
+    const response = await fetch(`http://localhost:3360/api/messages/user/${userId}`);
     const data = await response.json();
     setMessages(data);
   };
@@ -40,39 +43,30 @@ const Principal = () => {
   }, [selectedUser]);
 
   useEffect(() => {
-    const socket = new WebSocket('ws://localhost:3000');
+    const socket = new WebSocket('ws://localhost:3360');
     setWs(socket);
 
+    socket.onopen = () => {
+      console.log('WebSocket connection established');
+    };
+
     socket.onmessage = (event) => {
-      if (typeof event.data === 'string') {
-        // Si event.data es un string, se puede parsear directamente
-        try {
-          const message = JSON.parse(event.data);
-          if (message && message.remitente_id && message.destinatario_id) {
-            if (message.remitente_id === currentUser.id || message.destinatario_id === currentUser.id) {
-              setMessages((prevMessages) => [...prevMessages, message]);
-            }
-          }
-        } catch (e) {
-          console.error('Error processing message:', e);
-        }
+      const message = JSON.parse(event.data);
+      if (message.type === 'delete') {
+        setMessages((prevMessages) => prevMessages.filter((msg) => msg.id !== message.id));
       } else {
-        // Si event.data no es un string, usar FileReader para leerlo como texto
-        const reader = new FileReader();
-        reader.onload = () => {
-          try {
-            const message = JSON.parse(reader.result);
-            if (message && message.remitente_id && message.destinatario_id) {
-              if (message.remitente_id === currentUser.id || message.destinatario_id === currentUser.id) {
-                setMessages((prevMessages) => [...prevMessages, message]);
-              }
-            }
-          } catch (e) {
-            console.error('Error processing message:', e);
-          }
-        };
-        reader.readAsText(new Blob([event.data]));
+        if (message.destinatario_id === currentUser.id || message.remitente_id === currentUser.id) {
+          setMessages((prevMessages) => [...prevMessages, message]);
+        }
       }
+    };
+
+    socket.onclose = () => {
+      console.log('WebSocket connection closed');
+    };
+
+    socket.onerror = (error) => {
+      console.error('WebSocket error:', error);
     };
 
     return () => {
@@ -85,29 +79,47 @@ const Principal = () => {
   };
 
   const handleSendMessage = async () => {
-    if (inputValue.trim() !== '' && selectedUser && ws) {
+    if (inputValue.trim() !== '' && selectedUser) {
       const message = {
         remitente_id: currentUser.id,
         destinatario_id: selectedUser.id,
         mensaje: inputValue,
-        fecha_envio: new Date().toISOString().slice(0, 19).replace('T', ' '), // Formato adecuado para MySQL
+        fecha_envio: new Date().toISOString().slice(0, 19).replace('T', ' '),
       };
 
-      // Enviar el mensaje a través de WebSocket
-      ws.send(JSON.stringify(message));
+      try {
+        const response = await fetch('http://localhost:3360/api/messages', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify(message)
+        });
 
-      // Limpia el campo de entrada de texto
-      setInputValue('');
+        if (response.ok) {
+          console.log('Message sent to API');
+          setInputValue('');
+        } else {
+          console.error('Error sending message to API:', await response.text());
+        }
+      } catch (error) {
+        console.error('Error sending message to API:', error);
+      }
     }
   };
 
   const handleDeleteMessage = async (messageId) => {
+    if (!messageId) {
+      console.error('Message ID is invalid');
+      return;
+    }
+
     try {
-      const response = await fetch(`http://localhost:3000/api/messages/${messageId}`, {
+      const response = await fetch(`http://localhost:3360/api/messages/${messageId}`, {
         method: 'DELETE',
       });
       if (response.ok) {
-        setMessages((prevMessages) => prevMessages.filter((message) => message.id !== messageId));
+        console.log('Message deleted');
       } else {
         console.error('Error deleting message');
       }
@@ -118,10 +130,10 @@ const Principal = () => {
 
   const renderMessage = (message, index) => (
     message && (
-      <Box key={index} display="flex" flexDirection="column" alignItems="flex-start" mb={2}>
+      <Box key={index} display="flex" flexDirection="column" alignItems={message.remitente_id === currentUser.id ? 'flex-end' : 'flex-start'} mb={2}>
         <Typography variant="body2" color="textSecondary">{message.remitente_id === currentUser.id ? 'Tú' : selectedUser?.nombre}</Typography>
         <Box display="flex" alignItems="center">
-          <Box bgcolor="primary.main" color="white" p={2} borderRadius={2} maxWidth="80%">
+          <Box bgcolor={message.remitente_id === currentUser.id ? "primary.main" : "secondary.main"} color="white" p={2} borderRadius={2} maxWidth="80%">
             {message.mensaje}
           </Box>
           <Typography variant="body2" color="textSecondary" sx={{ ml: 1 }}>{new Date(message.fecha_envio).toLocaleTimeString()}</Typography>
@@ -137,6 +149,14 @@ const Principal = () => {
     const value = event.target.value;
     setSearchTerm(value);
     setFilteredUsers(users.filter(user => user.nombre.toLowerCase().includes(value.toLowerCase())));
+  };
+
+  const handleOpenPaymentModal = () => {
+    setIsPaymentModalOpen(true);
+  };
+
+  const handleClosePaymentModal = () => {
+    setIsPaymentModalOpen(false);
   };
 
   return (
@@ -169,8 +189,8 @@ const Principal = () => {
           <Divider />
         </List>
       </Box>
-      <Box sx={{ flexGrow: 1, display: 'flex', flexDirection: 'column' }}>
-        <Box sx={{ padding: 2, borderBottom: '1px solid #ccc', display: 'flex', alignItems:"center", justifyContent:"space-between" }}>
+      <Box sx={{ flexGrow: 1, display: 'flex', flexDirection: 'column', backgroundImage: `url(${chatBackground})`, backgroundSize: 'cover', backgroundPosition: 'center' }}>
+        <Box sx={{ padding: 2, borderBottom: '1px solid #ccc', display: 'flex', alignItems: "center", justifyContent: "space-between" }}>
           <Typography variant="h6">{selectedUser ? selectedUser.nombre : 'Seleccione un usuario'}</Typography>
           <IconButton>
             <MoreVert />
@@ -192,6 +212,36 @@ const Principal = () => {
           </Button>
         </Box>
       </Box>
+      <Box sx={{ width: '20%', padding: 2 }}>
+        <Paper elevation={3} sx={{ padding: 2, marginBottom: 2 }}>
+          <Typography variant="h6" gutterBottom>Información del Usuario</Typography>
+          <Typography variant="body1"><strong>Nombre:</strong> {currentUser?.nombre}</Typography>
+          <Typography variant="body1"><strong>Email:</strong> {currentUser?.email}</Typography>
+        </Paper>
+        <Paper elevation={3} sx={{ padding: 2, marginBottom: 2 }}>
+          <Typography variant="h6" gutterBottom>Contactos Recientes</Typography>
+          <List>
+            {filteredUsers.map((user) => (
+              <ListItem key={user.id}>
+                <ListItemAvatar>
+                  <Avatar src={MarkMesser} />
+                </ListItemAvatar>
+                <ListItemText primary={user.nombre} />
+              </ListItem>
+            ))}
+          </List>
+        </Paper>
+        <Paper elevation={3} sx={{ padding: 2 }}>
+          <Typography variant="h6" gutterBottom>Promociones</Typography>
+          <Typography variant="body1">
+            Aprovecha nuestras ofertas especiales en servicios de mensajería.
+          </Typography>
+        </Paper>
+      </Box>
+      <Button variant="contained" color="primary" onClick={handleOpenPaymentModal} sx={{ position: 'fixed', bottom: 16, right: 16 }}>
+        Realizar Pago 
+      </Button>
+      <PagoFicticioModal open={isPaymentModalOpen} handleClose={handleClosePaymentModal} userId={currentUser?.id} />
     </Container>
   );
 };
